motifdjango/requirements.txt
appdirs==1.4.0
awsebcli==3.9.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.3
blessed==1.9.5
botocore==1.5.6
bs4==0.0.1
cement==2.8.2
colorama==0.3.7
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1.1
django-el-pagination==3.0.1
django-libsass==0.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.1
libsass==0.12.3
MySQL-python==1.2.5
olefile==0.44
packaging==16.8
pathspec==0.5.0
Pillow==4.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
PyYAML==3.12
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
wcwidth==0.1.7
websocket-client==0.40.0
